use standard 1EU -> 4RF/FE conversion
This is such a standard that mods assume it and hardcode it, and people in the discord assumed it as well.  AE2 can be made to display its power consumption in RF, AE, or EU, and it hardcodes the 1:2:4 conversion rates, meaning that its EU output is wrong (it's 25% what you put in).